requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.17.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
breadability==0.1.20
Brotli==1.1.0
buildozer==1.5.0
bytesparse==1.1.0
cachetools==5.5.2
captum==0.8.0
certifi==2024.8.30
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.2
cmake==4.0.0
colorama==0.4.6
comm==0.2.2
comtypes==1.4.13
configs==3.0.3
contourpy==1.3.1
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.71.0
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dj-database-url==3.0.1
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
dlib==19.24.8
dnspython==2.7.0
docker==7.1.0
docopt==0.6.2
docutils==0.21.2
eventlet==0.39.1
executing==2.2.0
faiss-cpu==1.8.0
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpy==0.6.4
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
flask-cors==6.0.1
Flask-SocketIO==5.5.1
flatbuffers==24.3.25
fonttools==4.57.0
fpdf==1.7.2
fpdf2==2.8.5
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
googletrans==4.0.2
gradio==5.49.1
gradio_client==1.13.3
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.2.1
groovy==0.1.2
grpcio==1.73.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
h5py==3.12.1
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
imutils==0.5.4
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.4.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.7.4.5
keras==3.5.0
kiwisolver==1.4.8
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.3.82
langchain-huggingface==0.2.0
langchain-ollama==1.0.1
langchain-text-splitters==0.0.2
langsmith==0.6.2
libclang==18.1.1
libretranslatepy==2.1.1
lxml==5.4.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
mlflow==3.5.1
mlflow-skinny==3.5.1
mlflow-tracing==3.5.1
MouseInfo==0.1.3
moviepy==2.2.1
mpmath==1.3.0
multidict==6.6.4
multitasking==0.0.11
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.34.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.26.4
ollama==0.6.1
openai==2.15.0
opencv-python==4.10.0.84
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opt_einsum==3.4.0
optree==0.12.1
orjson==3.11.3
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.9
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==6.0.1
proglog==0.1.12
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyjokes==0.8.3
PyJWT==2.10.1
PyMsgBox==2.0.1
pyparsing==3.2.3
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
pysrt==1.1.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.13.0
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.9.11
requests==2.31.0
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.24.0
rsa==4.9.1
ruff==0.14.3
safehttpx==0.1.7
safetensors==0.5.3
scikeras==0.13.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==5.1.0
sh==2.2.2
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.1
sumy==0.11.0
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
tf_keras==2.19.0
threadpoolctl==3.6.0
timm==1.0.21
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.3
torch==2.9.0
torchvision==0.24.0
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.52.4
translate==3.6.1
typer==0.20.0
typer-slim==0.20.0
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.2.3
uuid_utils==0.13.0
uvicorn==0.34.3
views==0.3
virtualenv==20.27.1
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
widgetsnbextension==4.0.13
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.20.1
yfinance==0.2.55
youtube-dl==2021.12.17
youtube-transcript-api==1.0.3
zipp==3.23.0
zstandard==0.23.0
openai
python-dotenv
pydantic
fastapi
uvicorn
